Pulls beam graphics out into a file
Graphics.js is too large, let's split it up.

Forgive me, browsers, I'm making ya download another file.
