+++ index.html
@@ -19,6 +19,7 @@
 	<script type="module" src="src/util.js"></script>
 	<script type="module" src="src/file_utils.js"></script>
 	<script type="module" src="src/graphics.js"></script>
+	<script type="module" src="src/beam_graphics.js"></script>
 	<script type="module" src="src/player.js"></script>
 	<script type="module" src="src/view_state.js"></script>
 	<script type="module" src="src/menu.js"></script>
