+++ src/graphics.js
@@ -8,7 +8,7 @@
 import { _ } from "./singletons.js";
 import { to_radians, is_cheat_enabled, overridable_default } from "./util.js";
 
-const SHIP_Y = 0; // This might want to be imported from somewhere
+export const SHIP_Y = 0; // This might want to be imported from somewhere
 const PLANET_SCALE = 15;  // TODO: Noticing that differently sized planet sprites end up being the same screen-space size. Weird.
 const PLANET_Y = 0;  // TODO: Shots are still being drawn under planets for some reason
 
@@ -16,7 +16,7 @@ const STAR_Y = 0;
 
 const BG_LAYER = 0
 const SPOB_LAYER = 1
-const DEFAULT_LAYER = 2
+export const DEFAULT_LAYER = 2
 // const PLAYER_LAYER = 3
 
 let CAM_OFFSET_3DV = new BABYLON.Vector3(0, 0, 30);
@@ -436,53 +436,6 @@ export function material_from_skin(skin_data, dark){
     mat.custom_emissive_color = skin_data.custom_emissive_color;
   }
   return mat
-}
-
-export function get_beam(beam_info, length){
-  // TODO: Settle on a way to draw a nice beam
-  if(beam_info.type === "line"){
-    // TODO: This is totally broken. Why?
-    let line = BABYLON.MeshBuilder.CreateLines(
-      "line",
-      {
-        points: [
-          new BABYLON.Vector3(0,SHIP_Y,0),
-          new BABYLON.Vector3(0,SHIP_Y,length),
-        ]
-      },
-      _.scene,
-    );
-    line.color = new BABYLON.Color3(1, 1, 1);
-    return line;
-  }
-  if(beam_info.type === "tube"){
-    console.log(length);
-    let path = [
-      new BABYLON.Vector3(0,SHIP_Y,0),
-      new BABYLON.Vector3(length,SHIP_Y, 0),
-    ]
-
-    let beam = BABYLON.MeshBuilder.CreateTube(
-      "tube",
-      {
-        path: path,
-        radius: beam_info.radius || 1
-      },
-    );
-
-    let material = new BABYLON.StandardMaterial(_.scene);
-    material.alpha = 1;
-    material.diffuseColor = new BABYLON.Color3(
-      ...(beam_info.color || [1,1,1])
-    );
-    material.custom_emissive_color = beam_info.emissive_color;
-    beam.material = material;
-    beam.renderingGroupId = DEFAULT_LAYER;
-    return beam;
-  }
-  
-
-  // Alternate: I'd love to have particle-based beams too
 }
